applied mx eclipseformat to all Java files

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.graal.debug/src/com/oracle/graal/debug/Debug.java
@@ -243,7 +243,7 @@ public class Debug {
         public void increment() {
         }
 
-        public void add(long value) {
+        public void add(int value) {
         }
     };
