+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.graal.debug/src/com/oracle/graal/debug/DebugMetric.java
@@ -26,5 +26,5 @@ public interface DebugMetric {
 
     void increment();
 
-    void add(long value);
+    void add(int value);
 }
